Refactor: Standardize mediaId to String and update dependent screens
This commit standardizes the `mediaId` and `seriesId` types to `String` across various destination screens and ViewModels.

**Key Changes:**

- **`Destinations.kt`:**
    - `SeriesDetailsScreen`, `TopCastScreen`, `ReviewsScreen`, `SimilarMediaScreen`, `SeasonsScreen`, and `EpisodesScreen` now accept `mediaId` or `seriesId` as `String` instead of `Int`.
- **ViewModels:**
    - `SimilarMediaViewModel`, `SeriesDetailsViewModel`, and `ReviewsScreenViewModel` have been updated to parse `mediaId` or `seriesId` to `Int` when making API calls.
- **UI States:**
    - `SeriesDetailsUiState` now holds `seriesId` as a `String`.
    - `DetailsMovieUiState`'s `reviews` property type changed to `List<ReviewUiState>`.
- **Screens:**
    - **`SeeAllSimilarMediaScreen.kt`:** Updated navigation to `SeriesDetailsScreen` to pass `seriesId` as `String`.
    - **`SeriesDetailsScreen.kt`:**
        - Updated navigation to `SeriesDetailsScreen` (for similar series) to pass `seriesId` as `String`.
        - A helper function `toReviewScreenUiState` was added.
    - **`SimilarMoviesScreen.kt`:**
        - `SimilarMoviesSection` now accepts `List<Movie>` instead of `List<SimilarMovie>`.
        - `MovieCard` now uses `Movie` data class and navigates to `MovieDetailsScreen` with `movie.id`.
        - Removed `SimilarMovie` data class and preview code.
    - **`MovieDetailScreen.kt`:**
        - Updated navigation to `TopCastScreen`, `ReviewsScreen`, and `SimilarMediaScreen` to pass `mediaId` as `String`.
        - `SimilarMoviesSection` now correctly passes `uiState.similarMovies`.
        - Added `toReviewScreenUiStaten` helper function.
    - **`SearchScreen.kt`:** Updated navigation to `SeriesDetailsScreen` to pass `seriesId` as `String`.
- **Type Conversion:**
    - Explicit `toInt()` conversions are added in ViewModels where integer IDs are required for use case calls.

+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/ReviewsScreenViewModel.kt
@@ -28,7 +28,7 @@ class ReviewsScreenViewModel(
 
     private fun loadMovieReviews() {
         tryToExecute(
-            function = { movieDetailsUseCase.getMovieReviewsById(args.mediaId) },
+            function = { movieDetailsUseCase.getMovieReviewsById(args.mediaId.toInt()) },
             onSuccess = { reviews ->
                 updateState {
                     it.copy(reviews = reviews.map { review ->
@@ -42,7 +42,7 @@ class ReviewsScreenViewModel(
 
     private fun loadSeriesReviews() {
         tryToExecute(
-            function = { seriesDetailsUseCase.getSeriesReviewsById(args.mediaId) },
+            function = { seriesDetailsUseCase.getSeriesReviewsById(args.mediaId.toInt()) },
             onSuccess = { reviews ->
                 updateState {
                     it.copy(reviews = reviews.map { review ->
